add options for pinning 2nd hop relays for circuits
Added two new options:
  SecondHopHSMiddleNodes and SecondHopMiddleNodes

If SecondHopHSMiddleNodes is specified and the circuit being
built is a HS circuit, then Tor will choose one of the relays
from the given set randomly weighted by BW (using middle weights).
Otherwise, if SecondHopMiddleNodes is specified, then Tor chooses
the second hop relay similarly from this set of nodes. Tor will not
choose a node that exists in the ExcludeNodes set, and will not
choose the entry or exit that were already chosen for the circuit.

If neither option is given, Tor chooses middles normally.