added node signaling cell
Added a new option: SignalNodes

Whenever the Tor client finishes building a circuit, if the circuit
uses a relay in the SignalNodes set, then Tor will send a special
cell to that node to inform it that we chose it for the circuit.

This is only really useful if a single client and small set of
relays are running this code, since there are currently no
identifiers associated with the signal cell sent to the nodes.