send socks address on HS circuits through signal nodes
Whenever a BEGIN cell is sent through a relay in our SignalNodes
list, we send them the socks address associated with that stream.
The idea is that this address will contain the onion url that the
client is accessing. We hope to use this to signal to the middle
when our client is visiting facebook.