Fix cell events to correctly include relay commands
Relay commands should be included for both RELAY and RELAY_EARLY
cells.

For relay cells that originate at a relay, we were previously
sending them to the control prot after they had been crypted for
the next hop relay, and so the relay command was unreadable. We
now pass the pre-crypted relay header in these cases so that we
can preserve the relay command and related info.