change the way middles can be configured
Removed SecondHopHSMiddleNodes config option.
Added MiddleNodes config option.

If MiddleNodes is given, those nodes will be considered for
each middle position on a circuit.
If SecondHopMiddleNodes is given, those nodes will be considered
in addition to the MiddleNodes for the second hop middle position
in a circuit (the middle next to the guard node).